requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
django-extra-fields==0.9
django-filter==1.0.4
djangorestframework==3.6.3
mysqlclient==1.3.10
olefile==0.44
Pillow==4.2.1
pytz==2017.2
